fantasy/requirements.txt
yahoo_oauth
yahoo_fantasy_api
python-dotenv
schedule